% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add together two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number. I like \code{add()}, because it's great. It's obvious that
\code{add()} belongs to the same package as \code{\link[=set_mfl]{set_mfl()}}. For some irrelevant
information, see \code{vignette("rd-formatting", package = "roxygen2")}}

\item{y}{A number.}
}
\value{
A numeric vector.
}
\description{
\code{add()} adds together two numbers. Features of this function:
\itemize{
\item Smells nice
\item Has good vibes
}
}
\examples{
add(1, 1)
add(10, 1)

}
