% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkgs.playground-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{session_info}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sessioninfo}{\code{\link[sessioninfo]{session_info}}}
}}

