% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkgs.playground-package.R
\docType{package}
\name{rpkgs.playground-package}
\alias{rpkgs.playground}
\alias{rpkgs.playground-package}
\title{rpkgs.playground: Exercise for Learning Purpose}
\description{
An outcome of exercise for learning R package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lig3/rpkgs.playground}
  \item \url{https://lig3.github.io/rpkgs.playground/}
  \item Report bugs at \url{https://github.com/lig3/rpkgs.playground/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Gang Li \email{lig32@outlook.com}
  \item John Doe \email{johndoe@johndoe.com}
}

Other contributors:
\itemize{
  \item Company, Inc [copyright holder, funder]
}

}
\keyword{internal}
