% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mfl.R
\name{set_mfl}
\alias{set_mfl}
\title{Change my favorite letters}
\usage{
set_mfl(l = letters[24:26])
}
\arguments{
\item{l}{String to be set as \code{favorite_letters}, default to \code{letters[24:26]}}
}
\description{
Change my favorite letters
}
